requirements.txt
google-generativeai
pyinstaller
